refactor: move ui generic components into ui folder
