+++ src/pages/home/components/HeroSection/HeroSection.tsx
@@ -5,7 +5,7 @@ import heroBgDarkImage from "@public/hero-section-bg-dark-shape.svg"
 import heroBgLightImage from "@public/hero-section-bg-light-shape.svg"
 import spotifyListImage from "@public/spotify-list.svg"
 import Image from "next/image"
-import Chip from "./Chip"
+import Chip from "../../../../components/ui/Chip"
 import FloatingChips from "./FloatingChips"
 import HeroSectionRight from "./HeroSectionRight"
 import HeroSectionLeft from "./HeroSectionLeft"
