+++ src/pages/home/components/HeroSection/HeroSectionLeft.tsx
@@ -1,4 +1,4 @@
-import Button from "@/components/Button"
+import Button from "@/components/ui/Button"
 import { motion } from "framer-motion"
 import Link from "next/link"
